+++ website/src/constants.ts
@@ -1,4 +1,5 @@
 export const MAX_SPL_DECIMALS = 16;
+export const MAX_LAST_SPL_TOKENS = 9;
 
 export const NETWORK_SETTINGS_KEY = 'network';
 export const COMMITMENT_SETTINGS_KEY = 'commitment';