+++ CHANGELOG.md
@@ -13,7 +13,6 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Features
 
 * general: Show NFT info in wallet list.
-* general: Add SOL to SPL picker.
 * utilities: Add `SPL Utility` view to create/mint/burn SPL tokens.
 * utilities: Add `NFT Utility` view to create/mint/burn NFTs.
 
@@ -25,6 +24,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 * general: Add settings view.
 * general: Add wallet list menu-view.
 * general: Add menu to address tag to copy the address or see it in Solscan or Solana Explorer.
+* general: Add SOL to SPL picker.
+* general: Add last selected tokens to SPL selector.
 * config: Add support for different vs currencies.
 * config: Add support for different RPC endpoint.
 * config: Add support for different commitment levels.